ds4b_m1_apps/2_20216851-ZhenWang/ZhenWang_app/requirements.txt
altair
numpy
pandas
plotly
scikit-learn
shap
streamlit
streamlit_shap
xgboost
matplotlib
umap-learn
seaborn
pillow
